EXT-5389 Unreadable resident names with glow.
An alternative fix just for render-pipeline.
This is here instead of the trunk because it's cosmetically less good, and codewise a little weirder, and slightly (unmeasurably?) less performant.

But its advantage over the trunk fix is, it's still Z-buffered when FBOs are turned on.  Which we really care about on render-pipeline.

--HG--
extra : convert_revision : 83d7019fda75f5edef3900136e93c50ad83794b3