STORM-727 FIXED Don't disable Close and Back buttons in the Create Landmark panel, so you can go back if the landmark fails to load.
The bug was introduced in the fix of EXT-4700 (Creating a landmark brings up the Landmark sidepanel info twice).
To avoid reopening the "Create Landmark" panel, its Back and Close buttons were disabled.
However, the same fix removed the code for reopening the panel, so I can't see why we need to disable the buttons at all.