SH-557 FIX meshes sometimes don't load in a region
Found that we were storing an empty string as a getMesh capability in the case that
we checked for the new url before getting the list of caps from the server
after a teleport. Made sure we didn't overwrite the caps url until we have the
list of caps for the new region after teleport.

Will be post-reviewed by davep